feat(leaser): pass the release_from to all reases
Due to the fact that the lease lifecycle is strictly controlled by the leaser, on lease migration, the first lease is queried for its release, and that is passed to any subsequent lease migration.

misc: Had to split `try_close_protocol` into two functions to manage the increaded number of function parameters.