feat(lpp): On close all deposits, assert there are no loans.
We do assert because it is a precondition that the caller should have respected.
The caller, the protocol admin, i.e. the leaser contract, should have checked there are
no open leases before asking for Lpp close deposits.
An open loan would impede the collection of the required amounts in LPN to return to the deposit owners.