fix(leaser): Removing a lease cleans up the customer as well
The current logic for testing if there are open leases fails since it checks
whether there is any customer left in the leaser storage.

So far, the customer is kept with no leases when the last lease for a customer is being removed.
That has left a lot of dummy records, customers with empty set of leases.

In addition provided a cleaning migration logic to get rid of those records.