fix(lpp): Define and enforce a new invariant on the 'total interest due'
Due to rounding errors, the only computable data in LPP, the 'annual interest rate',
may have the 'total interest due' deviate a bit from the sum of the loan interests.
Re-calculating it on chain would be costly if there are a lot of open loans.
Hence we clear the deviation at the point the last loan is closed.